(wip) reorg dataset runner and tests and update tests
