+++ src/kimera_eval/__init__.py
@@ -2,6 +2,9 @@
 import kimera_eval.core.logger
 import kimera_eval.cli
 
+from kimera_eval.core.trajectory_metrics import *
+from kimera_eval.core.experiment_config import *
+
 import click
