+++ src/kimera_eval/config/NoVisualizer.flags
@@ -1,0 +1,12 @@
+--visualize=false
+--visualize_lmk_type=false
+--visualize_mesh=false
+--visualize_mesh_with_colored_polygon_clusters=false
+--visualize_point_cloud=false
+--visualize_convex_hull=false
+--visualize_plane_constraints=false
+--visualize_planes=false
+--visualize_plane_label=false
+--visualize_semantic_mesh=false
+--visualize_mesh_in_frustum=false
+--viz_type=2
