+++ src/kimera_eval/experiments/example_euroc.yaml
@@ -1,18 +1,17 @@
+---
 # Params used if running Kimera-VIO
-executable_path: '$HOME/Code/Kimera-VIO/build/stereoVIOEuroc'
-vocabulary_path: '$HOME/Code/Kimera-VIO/vocabulary/ORBvoc.yml'
-params_dir: '$HOME/Code/Kimera-VIO/params'
-
-# Params used for evaluation
-dataset_dir: '$HOME/Code/ROS/kimera_ws/src/Kimera-VIO-ROS/output_logs'
-results_dir: '$HOME/Code/Kimera-VIO-Evaluation/results'
-
+executable_path: $HOME/Code/Kimera-VIO/build/stereoVIOEuroc
+vocabulary_path: $HOME/Code/Kimera-VIO/vocabulary/ORBvoc.yml
+params_dir: $HOME/Code/Kimera-VIO/params
+extra_args: [--log_euroc_gt_data=true]
+pipelines:
+  - name: StereoEuroc
 datasets:
- - name: V1_01_easy
-   use_lcd: true
-   plot_vio_and_pgo: true
-   segments: [1]
-   discard_n_start_poses: 0
-   discard_n_end_poses: 0
-   initial_frame: 10
-   final_frame: 220
+  - name: V1_01_easy
+    initial_frame: 10
+    final_frame: 220
+    use_lcd: true
+    plot_vio_and_pgo: true
+    segments: [1]
+    discard_n_start_poses: 0
+    discard_n_end_poses: 0
