+++ tests/resources/test_experiments/test_euroc.yaml
@@ -1,0 +1,19 @@
+---
+# These three are only needed for running datasets, which isn't currently tested.
+executable_path: ''
+vocabulary_path: ''
+param_path: ''
+dataset_path: evaluation/tests/test_results
+pipelines:
+  - name: Euroc
+    param_name: Euroc
+datasets:
+  - name: V1_01_easy
+    use_lcd: true
+    plot_vio_and_pgo: true
+    segments: [1]
+    pipelines: [Euroc]
+    discard_n_start_poses: 0
+    discard_n_end_poses: 0
+    initial_frame: 10
+    final_frame: 3500
