+++ tests/resources/test_results/V1_01_easy/Euroc/traj_vio.csv
