--- tests/test_experiments/test_euroc.yaml
@@ -1,18 +1,0 @@
-# These three are only needed for running datasets, which isn't currently tested.
-executable_path: ''
-vocabulary_path: ''
-params_dir: ''
-
-results_dir: 'evaluation/tests/test_results'
-dataset_dir: 'evaluation/tests/test_results'
-
-datasets_to_run:
- - name: V1_01_easy
-   use_lcd: true
-   plot_vio_and_pgo: true
-   segments: [1]
-   pipelines: ['Euroc']
-   discard_n_start_poses: 0
-   discard_n_end_poses: 0
-   initial_frame: 10
-   final_frame: 3500
