+++ tox.ini
@@ -27,5 +27,5 @@ deps =
     pytest
     coverage
 commands =
-    python -m coverage run --source={envsitepackagesdir}/kimera_vio_evaluation -m pytest {posargs:tests}
-    python -m coverage report --show-missing --omit={envsitepackagesdir}/kimera_vio_evaluation/main.py
+    python -m coverage run --source={envsitepackagesdir}/kimera_eval -m pytest {posargs:tests}
+    python -m coverage report --show-missing --omit={envsitepackagesdir}/kimera_eval/cli.py
